Refactor check_neighbors method
The method was too arcane and unreadable with the previous approach. There is now a new method called check_cell_neighbors which takes a cell and just checks and counts them. The logic for wrapping is now in the Cell class which can now tell where its left, right, top, and bottom neighbors are